+++ src/models/danh-gia/roles/choHT.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
+import { arrange } from '../../../helpers/utilities';
 import { DanhGiaDocument } from '../danh-gia.entity';
 
 @Injectable()
@@ -54,7 +55,6 @@ export class ChoHieuTruongService {
                       }
                     : null,
                 tuanDG: all[i].tuanDG ? all[i].tuanDG : null,
-                chiTiet: all[i].chiTiet,
                 luotDG: all[i].chiTiet.length,
                 diemTB: all[i].chiTiet.length > 0 ? diem : 0,
             });
@@ -106,5 +106,114 @@ export class ChoHieuTruongService {
             luotDG: one.chiTiet.length,
             diemTB: one.chiTiet.length > 0 ? diem : 0,
         };
+    }
+
+    async getAll_forHT(tuan: string, lop = 'all') {
+        const query = { tuanDG: Object(tuan) };
+        if (Types.ObjectId.isValid(lop))
+            Object.assign(query, { lopHoc: Object(lop) });
+
+        const result = [];
+        const now = new Date().getTime();
+        const all = await this.model
+            .find(query)
+            .populate([
+                {
+                    path: 'giaoVien',
+                    select: 'hoTen',
+                },
+                { path: 'monHoc', select: 'tenMH' },
+                {
+                    path: 'lopHoc',
+                    select: ['maLH', 'hocSinh'],
+                },
+                {
+                    path: 'tuanDG',
+                    select: ['soTuan', 'ngayKetThuc'],
+                },
+            ])
+            .exec();
+
+        for (let i = 0; i < all.length; i++) {
+            let diem = 0,
+                temp = 0;
+            for (let j = 0; j < all[i].chiTiet.length; j++) {
+                temp += all[i].chiTiet[j].diemDG;
+            }
+            diem = temp / all[i].chiTiet.length;
+
+            result.push({
+                _id: all[i]._id,
+                tenDG: all[i].tenDG,
+                monHoc: all[i].monHoc?.tenMH,
+                lopHoc: {
+                    maLH: all[i].lopHoc?.maLH,
+                    siSo: all[i].lopHoc?.hocSinh.length,
+                },
+                giaoVien: all[i].giaoVien?.hoTen,
+                choGVCN: all[i].choGVCN,
+                daDuyet: all[i].daDuyet,
+                tuanDG: all[i].tuanDG?.soTuan,
+                luotDG: all[i].chiTiet.length,
+                diemTB: diem,
+                hetHan: arrange(all[i].tuanDG.ngayKetThuc).getTime() < now,
+            });
+        }
+        return result;
+    }
+
+    async getOne_forHT(dg: string) {
+        const one = await this.model
+            .findById(dg)
+            .populate([
+                {
+                    path: 'giaoVien',
+                    select: 'hoTen',
+                },
+                { path: 'monHoc', select: 'tenMH' },
+                { path: 'mauDG', select: 'tieuChi' },
+                {
+                    path: 'lopHoc',
+                    select: ['maLH', 'hocSinh'],
+                },
+                {
+                    path: 'tuanDG',
+                    select: 'soTuan',
+                },
+            ])
+            .exec();
+
+        let temp = 0,
+            diem = 0;
+        for (let i = 0; i < one.chiTiet.length; i++) {
+            temp += one.chiTiet[i].diemDG;
+        }
+        diem = temp / one.chiTiet.length;
+
+        return {
+            _id: dg,
+            tenDG: one.tenDG,
+            monHoc: one.monHoc?.tenMH,
+            giaoVien: one.giaoVien?.hoTen,
+            tieuChi: one.mauDG?.tieuChi,
+            choGVCN: one.choGVCN,
+            daDuyet: one.daDuyet,
+            lopHoc: {
+                maLH: one.lopHoc?.maLH,
+                siSo: one.lopHoc?.hocSinh.length,
+            },
+            tuanDG: one.tuanDG?.soTuan,
+            luotDG: one.chiTiet.length,
+            chiTiet: one.chiTiet,
+            diemDG: one.chiTiet.length > 0 ? diem : 0,
+        };
+    }
+
+    async approve(dg: string, tinhTrang = true) {
+        return await this.model.findByIdAndUpdate(
+            dg,
+            { $set: { daDuyet: tinhTrang } },
+            { new: true },
+        );
     }
 }