+++ src/users/hieu-truong/hieu-truong.controller.ts
@@ -1,4 +1,12 @@
-import { Controller, Get, Param, Query, UseGuards } from '@nestjs/common';
+import {
+    Body,
+    Controller,
+    Get,
+    Param,
+    Patch,
+    Query,
+    UseGuards,
+} from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { AuthGuard } from '../../auth.guard';
 import { HieuTruongService } from './hieu-truong.service';
@@ -9,11 +17,15 @@ import { HieuTruongService } from './hieu-truong.service';
 export class HieuTruongController {
     constructor(private service: HieuTruongService) {}
 
-    // @Get('danh-gia')
-    // @ApiOkResponse({ description: 'Trả về tất cả đánh giá trong CSDL' })
-    // async xemDanhGia() {
-    //     return await this.service.tatCa_DanhGia();
-    // }
+    @Get('danh-gia/theo')
+    async xemDanhGia(@Query('tuan') tuan: string, @Query('lop') lop: string) {
+        return await this.service.tatCa_DanhGia(tuan, lop);
+    }
+
+    @Get('danh-gia/:dg')
+    async xemMot_danhGia(@Param('dg') dg: string) {
+        return await this.service.motDanhGia(dg);
+    }
 
     @Get('danh-gia/bo-mon')
     async xemDanhGiaGVBM(@Query('gv') gv: string, @Query('lop') lop: string) {
@@ -33,5 +45,10 @@ export class HieuTruongController {
     @Get('danh-sach/giao-vien')
     async xemDS_giaoVien() {
         return await this.service.tatCa_GiaoVien();
+    }
+
+    @Patch('duyet-danh-gia/:dg')
+    async duyetDG(@Param('dg') dg: string, @Body() daDuyet?: boolean) {
+        return await this.service.duyetDG(dg, daDuyet);
     }
 }